Refactor kubetest2 gke deployer into different source files
