+++ kubetest2-gke/deployer/firewall.go
@@ -1,0 +1,148 @@
+/*
+Copyright 2020 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package deployer
+
+import (
+	"fmt"
+	"sort"
+	"strings"
+
+	"k8s.io/klog"
+
+	"sigs.k8s.io/kubetest2/pkg/exec"
+)
+
+func (d *deployer) ensureFirewall(project, cluster, network string) error {
+	if network == "default" {
+		return nil
+	}
+	firewall, err := d.getClusterFirewall(project, cluster)
+	if err != nil {
+		return fmt.Errorf("error getting unique firewall: %v", err)
+	}
+	if runWithNoOutput(exec.Command("gcloud", "compute", "firewall-rules", "describe", firewall,
+		"--project="+project,
+		"--format=value(name)")) == nil {
+		// Assume that if this unique firewall exists, it's good to go.
+		return nil
+	}
+	klog.V(1).Infof("Couldn't describe firewall '%s', assuming it doesn't exist and creating it", firewall)
+
+	tagOut, err := exec.Output(exec.Command("gcloud", "compute", "instances", "list",
+		"--project="+project,
+		"--filter=metadata.created-by:*"+d.instanceGroups[project][cluster][0].path,
+		"--limit=1",
+		"--format=get(tags.items)"))
+	if err != nil {
+		return fmt.Errorf("instances list failed: %s", execError(err))
+	}
+	tag := strings.TrimSpace(string(tagOut))
+	if tag == "" {
+		return fmt.Errorf("instances list returned no instances (or instance has no tags)")
+	}
+
+	if err := runWithOutput(exec.Command("gcloud", "compute", "firewall-rules", "create", firewall,
+		"--project="+project,
+		"--network="+network,
+		"--allow="+e2eAllow,
+		"--target-tags="+tag)); err != nil {
+		return fmt.Errorf("error creating e2e firewall: %v", err)
+	}
+	return nil
+}
+
+func (d *deployer) getClusterFirewall(project, cluster string) (string, error) {
+	if err := d.getInstanceGroups(); err != nil {
+		return "", err
+	}
+	// We want to ensure that there's an e2e-ports-* firewall rule
+	// that maps to the cluster nodes, but the target tag for the
+	// nodes can be slow to get. Use the hash from the lexically first
+	// node pool instead.
+	return "e2e-ports-" + d.instanceGroups[project][cluster][0].uniq, nil
+}
+
+// This function ensures that all firewall-rules are deleted from specific network.
+// We also want to keep in logs that there were some resources leaking.
+func (d *deployer) cleanupNetworkFirewalls(project, network string) (int, error) {
+	fws, err := exec.Output(exec.Command("gcloud", "compute", "firewall-rules", "list",
+		"--format=value(name)",
+		"--project="+project,
+		"--filter=network:"+network))
+	if err != nil {
+		return 0, fmt.Errorf("firewall rules list failed: %s", execError(err))
+	}
+	if len(fws) > 0 {
+		fwList := strings.Split(strings.TrimSpace(string(fws)), "\n")
+		klog.V(1).Infof("Network %s has %v undeleted firewall rules %v", network, len(fwList), fwList)
+		commandArgs := []string{"compute", "firewall-rules", "delete", "-q"}
+		commandArgs = append(commandArgs, fwList...)
+		commandArgs = append(commandArgs, "--project="+project)
+		errFirewall := runWithOutput(exec.Command("gcloud", commandArgs...))
+		if errFirewall != nil {
+			return 0, fmt.Errorf("error deleting firewall: %v", errFirewall)
+		}
+		return len(fwList), nil
+	}
+	return 0, nil
+}
+
+func (d *deployer) getInstanceGroups() error {
+	if d.instanceGroups != nil {
+		return nil
+	}
+
+	// Initialize project instance groups structure
+	d.instanceGroups = map[string]map[string][]*ig{}
+
+	location, err := d.location()
+	if err != nil {
+		return err
+	}
+
+	for _, project := range d.projects {
+		d.instanceGroups[project] = map[string][]*ig{}
+
+		for _, cluster := range d.projectClustersLayout[project] {
+			igs, err := exec.Output(exec.Command("gcloud", d.containerArgs("clusters", "describe", cluster,
+				"--format=value(instanceGroupUrls)",
+				"--project="+project,
+				location)...))
+			if err != nil {
+				return fmt.Errorf("instance group URL fetch failed: %s", execError(err))
+			}
+			igURLs := strings.Split(strings.TrimSpace(string(igs)), ";")
+			if len(igURLs) == 0 {
+				return fmt.Errorf("no instance group URLs returned by gcloud, output %q", string(igs))
+			}
+			sort.Strings(igURLs)
+
+			// Inialize cluster instance groups
+			d.instanceGroups[project][cluster] = make([]*ig, 0)
+
+			for _, igURL := range igURLs {
+				m := poolRe.FindStringSubmatch(igURL)
+				if len(m) == 0 {
+					return fmt.Errorf("instanceGroupUrl %q did not match regex %v", igURL, poolRe)
+				}
+				d.instanceGroups[project][cluster] = append(d.instanceGroups[project][cluster], &ig{path: m[0], zone: m[1], name: m[2], uniq: m[3]})
+			}
+		}
+	}
+
+	return nil
+}
